\name{data_cor}
\alias{data_cor}
\alias{data_pcor}
\alias{data_association}
\alias{high_val}
\alias{low_val}
\title{
Plotting pairwise linear and partial correlations.
}
\description{
The function \code{data_cor} takes a data frame and plots the pairwise Pearson's correlation coefficients of all continuous variables in the data.

The function \code{data_pcor} takes a data frame and plots the pairwise partial Pearson's correlation coefficients of all continuous variables in the data.

The function \code{data_association} takes a data frame and plots the pairwise association coefficients of all variables in the data. For contituous against continuous variables it shows  the absolute value of the Pearson's correlation coeficient, for categorical agaist categorical it shows Cramer's \eqn{\phi}, for continuous agaist categorigal it fit an analyis of variance model and reports the square root of the \eqn{R^2}.    

The functions \code{high_val} and \code{low_val} take the square matrix generated by the above thee functions and shows which paire-wise  have a value larger or smaller respectively,  than the value specified  by the argument by \code{val}.

}
\usage{
data_cor(data,  digits = 3, plot = TRUE, diag.off = TRUE,
        lower.tri.off = FALSE, method = c("square", "circle"),
        type = c("pearson", "kendall", "spearman"),
        outline.color = "gray", colors = c("blue", "white", "red"),
        legend.title = "Corr", title, ggtheme = theme_minimal(),
        tl.cex = 12, tl.col = "black", tl.srt = 45, lab = TRUE,
        lab_col = "black", lab_size = 3, circle.size = 20, seed = 123, 
        percentage,  print.info=TRUE)

data_pcor(data, digits = 3, plot = TRUE, diag.off = TRUE,
        lower.tri.off = FALSE, method = c("square", "circle"),
        outline.color = "gray", colors = c("blue", "white", "red"),
        legend.title = "Corr", title, ggtheme = theme_minimal(),
        tl.cex = 12, tl.col = "black", tl.srt = 45, lab = TRUE,
        lab_col = "black", lab_size = 3, circle.size = 20, seed = 123, 
        percentage, print.info=TRUE)

data_association(data, digits = 3, plot = TRUE, diag.off = TRUE, 
        lower.tri.off = FALSE, method = c("square", "circle"), 
        outline.color = "gray", colors = c("blue", "white", "red"), 
        legend.title = "Assoc", title, ggtheme = ggplot2::theme_minimal(), 
        tl.cex = 12, tl.col = "black", tl.srt = 45, lab = TRUE, 
        lab_col = "black", lab_size = 3, circle.size = 20, seed = 123, 
        percentage, print.info=TRUE )

high_val(table, val = 0.90, digits = 3, plot = FALSE, igraph = TRUE)

low_val(table, val = 0.05, digits = 3, plot = FALSE, igraph = TRUE)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{data}{a data frame}
   \item{table}{a correlation table obtained by \code{data_cor(,plot=FALSE)} or \code{data_pcor(,plot=FALSE)}}
  \item{digits}{the digits for printing the correlation coefficients}
  \item{plot}{whether to plot or not}
  \item{diag.off}{whether to show the diagonal ellements}
  \item{lower.tri.off}{whether to show the lower part of the matrix}
  \item{method}{plotting in \code{"square"} or \code{"cicle"}}
  \item{type}{type of correlation c("pearson", "kendall", "spearman")}
  \item{outline.color}{the outline colour}
  \item{colors}{the range of colours}
  \item{legend.title}{title for the legend}
  \item{title}{the main title}
  \item{ggtheme}{the theme for the plot, see package \pkg{ggthemes} for more themes}
  \item{tl.cex}{the text size for the marginal labels}
  \item{tl.col}{the colour of the he marginal labels}
  \item{tl.srt}{the angle of the text in the bottom labels of the table}
  \item{lab}{whether to show the correlation coefficients in the table}
  \item{lab_col}{the colour of the lettering of the correlation coefficients}
  \item{lab_size}{the size of the lettering of the correlation coefficients, increase (or decrease) if the defaul 3 is not appropriate}
  \item{circle.size}{the size of the circles, increase (or decrease) if the defaul 20 is not appropriate}
   \item{percentage}{this is for big data sets. if more tha a milion ony 10\% is plotted, if from 100.00 to a milion, 20\%, if 50.000 to 100.000, 50\% otherwise 100\% of the data.}
  \item{seed}{Setting a seed value for selection of the percantage of data (for big data sets)}
   \item{print.info}{whether to print infomation when cutting the data usinf  \code{data_cut()}}
   \item{val}{the theshold value so if tha actul value is  greater than \code{val} it will be reported in \code{high_val()}}
   \item{igraph}{if in \code{high_val()}, the option \code{plot=TRUE} is set, then there are two options for plotting; i)  \code{igraph::graph_from_data_frame()} with \code{igraph=TRUE} or ii) \code{networkD3::simpleNetwork()} with \code{igraph=FALSE} }
}

\value{
creates a correlation matrix plot.
}
\references{
Rigby, R. A., Stasinopoulos, D. M.,  Heller, G. Z.,  and De Bastiani, F. (2019)
\emph{Distributions for modeling location, scale, and shape: Using GAMLSS in R}, Chapman and Hall/CRC. \doi{10.1201/9780429298547}.

Stasinopoulos D. M., Rigby R.A., Heller G., Voudouris V., and De Bastiani F., (2017)
\emph{Flexible Regression and Smoothing: Using GAMLSS in R},  Chapman and Hall/CRC. \doi{10.1201/b21973}


Stasinopoulos, M.D., Kneib, T., Klein, N., Mayr, A. and Heller, G.Z., (2024). \emph{Generalized Additive Models for Location, Scale and Shape: A Distributional Regression Approach, with Applications} (Vol. \bold{56}). Cambridge University Press.

(see also \url{https://www.gamlss.com/}).
}
\author{Mikis Stasinopoulos}


\seealso{mcor}
\examples{
data_cor(rent99)
Pearson.cor <-  data_cor(rent99, plot=FALSE)

data_pcor(rent99)
partial.cor  <-  data_cor(rent99, plot=FALSE) 


high_val(partial.cor, val=0.5)
high_val(Pearson.cor, val=0.5)
}

\keyword{regression}

